Updated HelperFunction and Created new Components
