+++ package.json
@@ -9,7 +9,7 @@
     "axios": "^1.4.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router": "^6.11.2",
+    "react-router-dom": "^6.11.2",
     "react-scripts": "5.0.1",
     "react-youtube": "^10.1.0",
     "web-vitals": "^2.1.4"
