+++ src/Components/About.js
@@ -1,0 +1,7 @@
+export default function About() {
+  return (
+    <div>
+      <p>About Page</p>
+    </div>
+  );
+}
