+++ src/Components/Cards.js
@@ -1,1 +1,22 @@
-export default function Cards() {}
+import fetchApi from "./api/fetch";
+import { useEffect, useState } from "react";
+
+export default function Cards() {
+  const [videoData, setVideoData] = useState([]);
+
+  async function fetchStandardResults() {
+    try {
+      let result = await fetchApi({
+        method: "get",
+      });
+      console.log(result.data);
+      setVideoData(result.data);
+    } catch (error) {
+      console.log(error);
+    }
+  }
+  useEffect(() => {
+    fetchStandardResults();
+    console.log(videoData?.items);
+  }, []);
+}
