+++ src/Components/Channel.js
@@ -1,0 +1,3 @@
+import React, { useEffect, useState } from "react";
+import fetchApi from "./api/fetch";
+export default function Channel() {}
